requirements.txt
pytest
airtest
pocoui
allure-pytest
